Move IndexViewModel to partial class
