+++ src/Traces.Web/Pages/Index.razor
@@ -1,7 +1,4 @@
 ﻿@page "/"
-@using Traces.Web.ViewModels
-@inherits IndexViewModel
-
 <div>
     <Card>
         <CardBody>
